refactor(pi-codex-conversion): own voice preflight in voice
