+++ packages/pi-codex-conversion/src/voice/AGENTS.md
@@ -1,4 +1,5 @@
 - `controller.ts` owns the public facade, teardown, mute, and Pi event bridge; `controller-start.ts` owns startup/auth/state transitions; `controller-sessions.ts` loads modes and guards cancellation across imports; `controller-support.ts` owns shared state/presentation helpers.
+- `delegation-preflight.ts` owns prompt/tool prewarm identity plus commit/rollback before Pi receives an idle voice delegation.
 - Under `conversation/`, `session.ts` owns V3 sequencing, `call-setup.ts` HTTP setup, `handoff.ts` delegation output, and `wire.ts` validation. `auth.ts`, `dictation/`, and `session-messages.ts` retain their boundaries.
 - `helper.ts` owns the process; `helper-protocol.ts` owns JSONL framing and validation. LAN browser transport, ownership, and decoding stay in `browser-connections.ts`, `browser-session.ts`, and `browser-wire.ts`.
 - `controls.ts` owns start/stop/setup policy; settings commands and shortcuts route through it into the controller.
